+++ pennylane/_qubit_device.py
@@ -17,10 +17,8 @@ This module contains the :class:`QubitDevice` abstract base class.
 
 # For now, arguments may be different from the signatures provided in Device
 # e.g. instead of expval(self, observable, wires, par) have expval(self, observable)
-# pylint: disable=arguments-differ, abstract-method, no-value-for-parameter
+# pylint: disable=arguments-differ, abstract-method, no-value-for-parameter,too-many-instance-attributes
 
-import itertools
-from collections import OrderedDict
 import numpy as np
 
 from pennylane.operation import Sample, Variance, Expectation, Probability
@@ -43,6 +41,21 @@ class QubitDevice(Device):
         self.analytic = analytic
 
         self._state = None
+        self._rotated_prob = None
+        self._wires_used = None
+        self._memory = None
+        self._samples = None
+
+
+    def reset(self):
+        """Reset the backend state.
+
+        After the reset, the backend should be as if it was just constructed.
+        Most importantly the quantum state is reset to its initial value.
+        """
+        self._state = None
+        self._rotated_prob = None
+        self._wires_used = None
         self._memory = None
         self._samples = None
 
@@ -74,8 +87,6 @@ class QubitDevice(Device):
         self._parameters = {}
         self._parameters.update(parameters)
 
-        results = []
-
         with self.execution_context():
             self.pre_apply()
 
@@ -83,26 +94,13 @@ class QubitDevice(Device):
                 # Pass instances directly
                 self.apply(operation)
 
+            self.rotate_basis(observables)
+
             self.post_apply()
 
             self.pre_measure()
 
-            for obs in observables:
-                # Pass instances directly
-                if obs.return_type is Expectation:
-                    results.append(self.expval(obs))
-
-                elif obs.return_type is Variance:
-                    results.append(self.var(obs))
-
-                elif obs.return_type is Sample:
-                    results.append(np.array(self.sample(obs)))
-
-                elif obs.return_type is Probability:
-                    results.append(list(self.probability(wires=obs.wires).values()))
-
-                elif obs.return_type is not None:
-                    raise QuantumFunctionError("Unsupported return type specified for observable {}".format(obs.name))
+            results = self.statistics(observables)
 
             self.post_measure()
 
@@ -112,33 +110,136 @@ class QubitDevice(Device):
 
             # Ensures that a combination with sample does not put
             # expvals and vars in superfluous arrays
-            if all(obs.return_type is Sample for obs in observables):
-                return self._asarray(results)
-            if any(obs.return_type is Sample for obs in observables):
+            sample_return_types = (obs.return_type is Sample for obs in observables)
+            if any(sample_return_types) and not all(sample_return_types):
                 return self._asarray(results, dtype="object")
 
             return self._asarray(results)
 
-    def rotate_basis(self, observable):
+    def statistics(self, observables):
+        """Process measurement results from circuit execution and return statistics.
+
+        This includes returning expectation values, variance, samples and probabilities.
+
+        Args:
+            observables (Union[:class:`Observable`, List[:class:`Observable`]]): the number of basis states to sample from
+
+        Raises:
+            QuantumFunctionError: if the value of :attr:`~.Observable.return_type` is not supported
+
+        Returns:
+            Union[float, List[float]]: the corresponding statistics
+        """
+        results = []
+
+        for obs in observables:
+            # Pass instances directly
+            if obs.return_type is Expectation:
+                results.append(self.expval(obs))
+
+            elif obs.return_type is Variance:
+                results.append(self.var(obs))
+
+            elif obs.return_type is Sample:
+                results.append(np.array(self.sample(obs)))
+
+            elif obs.return_type is Probability:
+                results.append(self.probability(wires=obs.wires))
+
+            elif obs.return_type is not None:
+                raise QuantumFunctionError("Unsupported return type specified for observable {}".format(obs.name))
+
+        return results
+
+    def rotate_basis(self, obs_queue):
         """Rotates the specified wires such that they
         are in the eigenbasis of the provided observable.
 
         Args:
-            observable (Observable): the observable that is to be
-                measured
+            observables (List[:class:`Observable`]): the observable queue
         """
-        for diag_gate in observable.diagonalizing_gates():
-            self.apply(diag_gate)
+        self._memory = False
+
+        wires = []
+
+        for observable in obs_queue:
+            if hasattr(observable, "return_type") and observable.return_type == Sample:
+                self._memory = True  # make sure to return samples
+
+            for diag_gate in observable.diagonalizing_gates():
+                self.apply(diag_gate)
+
+            for wire in observable.wires:
+                if isinstance(wire, int):
+                    wires.append(wire)
+                else:
+                    wires.extend(wire)
+
+        # Store the wires used by the observables such that
+        # an Identity is considered on the remaining wires
+        self._wires_used = wires
+        self._rotated_prob = self.probability(wires)
+
+    def generate_samples(self):
+        """Generate computational basis samples based on the current state.
+
+        If the device contains a sample return type, or the
+        device is running in non-analytic mode, ``dev.shots`` number of
+        computational basis samples are generated and stored within
+        the :attr:`~._samples` attribute.
+
+        .. warning:: This method should be overwritten on devices that
+            generate their own computational basis samples.
+        """
+        number_of_states = 2**len(self._wires_used)
+        samples = self.sample_basis_states(number_of_states, self._rotated_prob)
+        self._samples = QubitDevice.states_to_binary(samples, number_of_states)
+
+    def sample_basis_states(self, number_of_states, state_probability):
+        """Sample from the computational basis states based on the state
+        probability.
+
+        This is an auxiliary method to the generate_samples method.
+
+        Args:
+            number_of_states (int): the number of basis states to sample from
+
+        Returns:
+            List[int]: the sampled basis states
+        """
+        basis_states = np.arange(number_of_states)
+        return np.random.choice(basis_states, self.shots, p=state_probability)
+
+    @staticmethod
+    def states_to_binary(samples, number_of_states):
+        """Convert basis states from base 10 to binary representation.
+
+        This is an auxiliary method to the generate_samples method.
+
+        Args:
+            samples (List[int]): samples of basis states in base 10 representation
+            number_of_states (int): the number of basis states to sample from
+
+        Returns:
+            List[int]: basis states in binary representation
+        """
+        powers_of_two = (1 << np.arange(number_of_states))
+        states_sampled_base_ten = samples[:, None] & powers_of_two
+        return (states_sampled_base_ten > 0).astype(int)
 
     def expval(self, observable):
         wires = observable.wires
-        self.rotate_basis(observable)
+
+        # TODO: remove the following part as we can assume that
+        # rotate_basis has already been called. Doing so will, however break
+        # unit tests directly calling on expval()
+        if self._rotated_prob is None:
+            self.rotate_basis([observable])
 
         if self.analytic:
             # exact expectation value
             eigvals = observable.eigvals
-            prob = np.abs(self._state ** 2)
-            prob = self.marginal_prob(prob, wires=wires)
+            prob = self.probability(wires=wires)
             return (eigvals @ prob).real
 
         # estimate the ev
@@ -146,21 +247,85 @@ class QubitDevice(Device):
 
     def var(self, observable):
         wires = observable.wires
-        self.rotate_basis(observable)
+
+        # TODO: remove the following part as we can assume that
+        # rotate_basis has already been called. Doing so will, however break
+        # unit tests directly calling on var()
+        if self._rotated_prob is None:
+            self.rotate_basis([observable])
 
         if self.analytic:
             # exact variance value
             eigvals = observable.eigvals
-            prob = np.abs(self._state ** 2)
-            prob = self.marginal_prob(prob, wires=wires)
+            prob = self.probability(wires=wires)
             return (eigvals ** 2) @ prob - (eigvals @ prob).real ** 2
 
+        # estimate the variance
         return np.var(self.sample(observable))
 
+    def sample(self, observable):
+        wires = observable.wires
+        name = observable.name
 
-    def probability(self, wires=None, values_only=False):
+        # TODO: remove the following part as we can assume that
+        # rotate_basis has already been called. Doing so will, however break
+        # unit tests directly calling on sample()
+        if self._rotated_prob is None:
+            observable.return_type = Sample
+            self.rotate_basis([observable])
+
+            if self._memory or (not self.analytic):
+                self.generate_samples()
+
+        if isinstance(name, str) and name in {"PauliX", "PauliY", "PauliZ", "Hadamard"}:
+            return self.pauli_eigvals_as_samples(wires)
+
+        # Need to post-process the samples using the observables.
+        return self.custom_eigvals_as_samples(wires, observable.eigvals)
+
+    def pauli_eigvals_as_samples(self, wires):
+        """Process samples for observables with eigenvalues {1, -1}.
+
+        This method should be called for observables having eigenvalues {1, -1},
+        such that the post-processing step is known.
+
+        Args:
+            wires (Sequence[int]): Sequence of wires to return
+
+        Returns:
+            Sequence[int]: standard eigenvalues
+        """
+        return 1 - 2 * self._samples[:, wires[0]]
+
+
+    def custom_eigvals_as_samples(self, wires, eigenvalues):
+        """Replace the basis state in the computational basis with the correct eigenvalue.
+
+        Need to post-process the samples using the observables.
+        Extract only the columns of the basis samples required based on ``wires``.
+
+        Args:
+            wires (Sequence[int]): Sequence of wires to return
+            eigenvalues (Sequence[complex]): eigenvalues of the observable
+
+        Returns:
+            Sequence[complex]: the sampled eigenvalues of the observable
+        """
+        wires = np.hstack(wires)
+        samples = self._samples[:, np.array(wires)]
+        unraveled_indices = [2] * len(wires)
+        indices = np.ravel_multi_index(samples.T, unraveled_indices)
+        return eigenvalues[indices]
+
+    def probability(self, wires=None):
         """Return the (marginal) probability of each computational basis
         state from the last run of the device.
+
+        If no wires are specified, then all the basis states representable by
+        the device are considered and no marginalization takes place.
+
+        .. warning:: This method will have to be redefined for hardware devices, since it uses
+            the ``device._state`` attribute. This attribute might not be available for such devices.
 
         Args:
             wires (Sequence[int]): Sequence of wires to return
@@ -168,27 +333,21 @@ class QubitDevice(Device):
                 are traced out of the system.
 
         Returns:
-            OrderedDict[tuple, float]: Dictionary mapping a tuple representing the state
-            to the resulting probability. The dictionary should be sorted such that the
-            state tuples are in lexicographical order.
+            List[float]: list of the probabilities
         """
         if self._state is None:
             return None
 
         wires = wires or range(self.num_wires)
-
         prob = self.marginal_prob(np.abs(self._state)**2, wires)
-
-        if values_only:
-            return prob
-
-        basis_states = itertools.product(range(2), repeat=len(wires))
-        return OrderedDict(zip(basis_states, prob))
+        return prob
 
     def marginal_prob(self, prob, wires=None):
-        """Return the marginal probability of each computational basis
-        state from the last run of the device.
+        """Return the marginal probability of the computational basis
+        states by summing the probabiliites on the non-specified wires.
 
+        If no wires are specified, then all the basis states representable by
+        the device are considered and no marginalization takes place.
 
         Args:
             prob: The probabilities to return the marginal probabilities
@@ -196,7 +355,6 @@ class QubitDevice(Device):
             wires (Sequence[int]): Sequence of wires to return
                 marginal probabilities for. Wires not provided
                 are traced out of the system.
-
 
         Returns:
             list[float]: List of the resulting marginal probabilities.